Made control board all through-hole.
Reasoning: For JLC production it usually ends up cheaper to not provide user panelized boards with "two designs". By making only one board SMD, we can de-panelize and only do PCBA on the main board - this is about 20% cheaper (for 10 pairs of boards). The tradeoff is we have to solder 8 extra THT resistors.